+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Release Versions
 ## Upcoming changes
 
 - fix: raise exception when accessing an empty parameter value in python (#260)
+- feat!: rename set_parameter_value to set_parameter and add strict parameter map (#259)
 
 ## 9.3.1